+++ packages/orq-rc/src/orq_ai_sdk/models/fileuploadop.py
@@ -94,5 +94,5 @@ class FileUploadResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = dateutil.parser.isoparse("2025-03-04T21:40:20.059Z")
+    created: Optional[datetime] = dateutil.parser.isoparse("2025-03-04T21:45:16.894Z")
     r"""The date and time the resource was created"""