+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasetop.py
@@ -79,5 +79,5 @@ class RetrieveDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = dateutil.parser.isoparse("2025-03-04T21:40:18.702Z")
+    updated: Optional[datetime] = dateutil.parser.isoparse("2025-03-04T21:45:15.768Z")
     r"""The date and time the resource was last updated"""